Makes the main CMakeLists.txt more portable.
The POSIX version was previously only set in files. Having it set at the
root level makes more sense, especially considering that the previous
method could make successful compilation of the project dependent on the
order in which files are compiled. Which obviously not acceptable.
Flags were also set in a way that made them potentially only compatible
with GNUCC. This is now fixed (the optimization flag remains GNUCC only,
but a check for which compiler is used precedes it).